class structure created with indeed search working when chosen from a dropdown provided via enquirer. New jobsearching platforms can now be added into this class structure
